Move context impl to separate file
